+++ CHANGELOG.md
@@ -12,13 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Added**
 
 - added `ray-on-eks`, and `manifests/ray-on-eks` manifests
-- Added a `sagemaker-model-monitoring-module` module with an example of data quality monitoring of a SageMaker Endpoint.
+- Added a `sagemaker-model-monitoring-module` module with an example of data quality and model quality monitoring of a SageMaker Endpoint.
 - Added an option to enable data capture in the `sagemaker-endpoint-module`.
 
 ### **Changed**
 - remove explicit module manifest account/region mappings from `fmops-qna-rag`
 - moved CI/CD infra to separate repository and added self mutation pipeline to provision infra for module `sagemaker-templates-service-catalog`
 - changed ECR encryption to KMS_MANAGED
+- changed encryption for each bucket to KMS_MANAGED
 
 ## v1.2.0